refactor: rename service implementation classes and update package structure
