+++ src/main/java/dev/araopj/hrplatformapi/employee/service/impl/EmploymentInformationSalaryOverrideServiceImp.java
@@ -1,10 +1,11 @@
-package dev.araopj.hrplatformapi.employee.service;
+package dev.araopj.hrplatformapi.employee.service.impl;
 
 import dev.araopj.hrplatformapi.employee.dto.request.EmploymentInformationSalaryOverrideRequest;
 import dev.araopj.hrplatformapi.employee.dto.response.EmploymentInformationSalaryOverrideResponse;
 import dev.araopj.hrplatformapi.employee.model.EmploymentInformationSalaryOverride;
 import dev.araopj.hrplatformapi.employee.repository.EmploymentInformationRepository;
 import dev.araopj.hrplatformapi.employee.repository.EmploymentInformationSalaryOverrideRepository;
+import dev.araopj.hrplatformapi.employee.service.IEmploymentInformationSalaryOverrideService;
 import dev.araopj.hrplatformapi.exception.NotFoundException;
 import dev.araopj.hrplatformapi.utils.*;
 import lombok.RequiredArgsConstructor;
