+++ src/main/java/dev/araopj/hrplatformapi/employee/service/impl/PositionServiceImp.java
@@ -1,9 +1,10 @@
-package dev.araopj.hrplatformapi.employee.service;
+package dev.araopj.hrplatformapi.employee.service.impl;
 
 import dev.araopj.hrplatformapi.employee.dto.request.PositionRequest;
 import dev.araopj.hrplatformapi.employee.dto.response.PositionResponse;
 import dev.araopj.hrplatformapi.employee.repository.EmploymentInformationRepository;
 import dev.araopj.hrplatformapi.employee.repository.PositionRepository;
+import dev.araopj.hrplatformapi.employee.service.IPositionService;
 import dev.araopj.hrplatformapi.exception.NotFoundException;
 import dev.araopj.hrplatformapi.utils.AuditUtil;
 import dev.araopj.hrplatformapi.utils.DiffUtil;
